feat!(math): replace `add` function with more flexible `sum`
The `add` function has been removed in favor of a new `sum` function, which accepts an arbitrary number of arguments. This change allows for more flexible operations, enabling the summation of more than just two numbers. The `sum` function uses the `reduce` method to accumulate the total sum of all the provided numbers, enhancing the math package's capabilities significantly.

This is a breaking change as the `add` function is no longer available and has been replaced with the `sum` function, which operates differently by taking any number of parameters. Users relying on the `add` function will need to update their code to use the new `sum` function.