feat(functions): rename `running` event to `start` in `createFsm`
The `running` event in the `createFsm` function has been renamed to `start`. This change was made to better reflect the initial state of the finite state machine (FSM) when it begins its execution process. The update includes both the dispatching of the event in the FSM class and the adjustment in related unit tests to listen for the `start` event instead of `running`. This should make the FSM's lifecycle more intuitive to understand and use.